requirements.txt
nose
sphinx
tinycards==0.23
python-dotenv
pyyaml
gitpython
requests